* Revert didn't unschedule dir-props. * SVN::XD:Editor didn't schedule for deletion in merge mode. * Don't use DelayEditor in svk st, since it requires the callback   to be called with the StatusEditor. * Let close_directory handle conflict entries that is not opened. * Let revert remove files schedule for add.
git-svn-id: file:///home/chmrr/work/bps/complete-history/svk/orig-svn/trunk@176 4ccb8a5b-8ae0-0310-9b02-86531366991d

+++ lib/SVK/XD.pm
@@ -321,7 +321,7 @@ sub do_delete {
 	 }, @paths) if @paths;
 
     for (@deleted) {
-	print "D  $_\n";
+	print "D  $_\n" unless $arg{quiet};
 	$self->{checkout}->store ($_, {'.schedule' => 'delete'});
     }
 
@@ -400,9 +400,12 @@ sub do_revert {
     };
 
     my $unschedule = sub {
+	my $sche = $self->{checkout}->get ($_[1])->{'.schedule'};
 	$self->{checkout}->store ($_[1],
 				  {'.schedule' => undef,
 				   '.newprop' => undef});
+	-d $_[1] ? rmtree ([$_[1]]) : unlink($_[1])
+	    if $sche eq 'add';
 	print loc("Reverted %1\n", $_[1]);
     };
 
@@ -1035,7 +1038,15 @@ sub delete_entry {
     $self->{get_copath}($copath);
     # XXX: check if everyone under $path is sane for delete";
     return if $self->{check_only};
-    -d $copath ? rmtree ([$copath]) : unlink($copath);
+    if ($self->{update}) {
+	-d $copath ? rmtree ([$copath]) : unlink($copath);
+    }
+    else {
+	$self->{xd}->do_delete (%$self,
+				path => "$self->{anchor}/$path",
+				copath => $copath,
+				quiet => 1);
+    }
 }
 
 sub close_directory {
